refactor : 표준 Response DTO 패키지 위치 변경
