+++ src/main/java/com/genai/tcoop/exception/GlobalControllerAdvice.java
@@ -1,6 +1,6 @@
 package com.genai.tcoop.exception;
 
-import com.genai.tcoop.response.Response;
+import com.genai.tcoop.model.dto.response.Response;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
